Created a tune_params cross validation
We are running into a problem to tune params where the features need to be regenerated for each new params. This means that it can take up to 30 mins to test 1 set of params. I have made it so that with this method only the TF-IDF is regenerated, which means the time is cut by about 50% per each fold.